Moved tagger implemenation back into same constrution as the backwards vacuum
