+++ compact/far_backward.xml
@@ -4,7 +4,6 @@
   <include ref="far_backward/definitions.xml"/>
   <include ref="far_backward/magnets.xml"/>
   <include ref="far_backward/vacuum.xml"/>
-  <include ref="far_backward/taggers.xml"/>
   <!-- include ref="far_backward/lumi.xml"/ -->
   <include ref="far_backward/beamline_hadron.xml"/>
