+++ compact/far_backward/vacuum.xml
@@ -2,6 +2,8 @@
 <lccdd>
 
   <define>
+ 
+   <constant name="tracker_pixel_size" value="55*um"/>
 
   </define>
 
@@ -24,7 +26,7 @@
       <comment> Vacuum volume joining with lumi </comment>
       <exitdim    x="Exit_Width"   y="Exit_Height"  lumiZ="LumiWin_Z" lumiR="LumiWin_R" maxTheta="Exit_Theta"/>
 
-      <comment> Vacuum volume extention around taggers </comment>
+      <comment> Tagger stations and extended vacuum volume </comment>
       <module id="1"
         modname="Tagger1"
         theta_min="Tagger1_Min_Theta"
@@ -33,6 +35,11 @@
         extend_vacuum="true"
         >
         <dimensions x="Tagger1_Width/2" y="Tagger1_Height/2" z="Tagger1_Length"/>
+        <trackLayer  id="0" type="timepix" z="0*mm"   sensor_thickness="400*um"/>
+        <trackLayer  id="1" type="timepix" z="300*mm" sensor_thickness="400*um"/>
+        <trackLayer  id="2" type="timepix" z="600*mm" sensor_thickness="400*um"/>
+        <trackLayer  id="3" type="timepix" z="900*mm" sensor_thickness="400*um"/>
+        <!--windowLayer id="1" type="window"  z="0*mm"   thickness="1*mm" material="Copper"/-->
       </module>
 
       <module id="2"
@@ -43,10 +50,24 @@
         extend_vacuum="true"
         >
         <dimensions x="Tagger2_Width/2" y="Tagger2_Height/2" z="Tagger2_Length"/>
+        <trackLayer  id="0" type="timepix" z="0*mm"   sensor_thickness="400*um"/>
+        <trackLayer  id="1" type="timepix" z="300*mm" sensor_thickness="400*um"/>
+        <trackLayer  id="2" type="timepix" z="600*mm" sensor_thickness="400*um"/>
+        <trackLayer  id="3" type="timepix" z="900*mm" sensor_thickness="400*um"/>
+        <!--windowLayer id="1" type="window"  z="0*mm"   thickness="1*mm" material="Copper"/-->
       </module>
 
     </detector>
 
   </detectors>
+
+  <readouts>
+
+    <readout name="TaggerTrackerHits">
+      <segmentation type="CartesianGridXY" grid_size_x="tracker_pixel_size" grid_size_y="tracker_pixel_size" />
+      <id>system:8,module:2,layer:4,sensor:4,x:32:-16,y:-16</id>
+    </readout>
+
+  </readouts>
 
 </lccdd>
